Fixed compile warning

+++ src/paths/catmullromspline.cpp
@@ -74,10 +74,10 @@ void CatmullRomSpline::transform(const Matrix& m) {
 
 unsigned int CatmullRomSpline::segmentBegin(const double t) const {
     // TODO: Test!
-    int segments = points_num - 3;
+    unsigned int segments = points_num - 3;
     if (t == 1) return segments + 1;
     if (t == 0) return 1;
-    return floor(t * segments);
+    return (unsigned int)(floor(t * segments));
 }
 
 double CatmullRomSpline::adjustT(const double t) const {
